refactor: move components outside app directory
- Move app/components → components/
- Move app/contexts/AuthContext → components/providers/AuthContext
- Separate routing (app/) from shared components
- Follow Next.js 15 App Router best practices
