+++ components/editor/EditorWithToolbar.tsx
