+++ components/editor/extensions/ResizableImage.ts
