+++ components/editor/extensions/ResizableImageComponent.tsx
