+++ components/editor/index.ts
