+++ components/landing/HeroSection.tsx
