+++ components/landing/HowItWorksSection.tsx
